% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCClocation2FIPS.R
\name{FCClocation2FIPS}
\alias{FCClocation2FIPS}
\title{Convert FCC location to FIPS code}
\usage{
FCClocation2FIPS(place_id, lat, lon)
}
\arguments{
\item{place_id}{some unique identifier for the lat lon}

\item{lat}{the latitude}

\item{lon}{the longitude}
}
\value{
data.table
}
\description{
Convert FCC location to FIPS code
}
\examples{
sdalr::FCClocation2FIPS("VTRC", lat=38.880807, lon=-77.11577)
}
