% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCClocation2FIPS.R
\name{FCClocations2FIPS}
\alias{FCClocations2FIPS}
\title{Convert multiple FCC locations to FIPS codes}
\usage{
FCClocations2FIPS(place_idCol, latCol, lonCol)
}
\arguments{
\item{place_idCol}{vector of unique identifiers}

\item{latCol}{vector of latitudes}

\item{lonCol}{vector of longitudes}
}
\value{
data.table
}
\description{
Convert multiple FCC locations to FIPS codes
}
\examples{
sdalr::FCClocations2FIPS(place_idCol = c("VTRC", "VT-NVC"),
                         latCol = c(38.880807, 38.8968325),
                         lonCol = c(-77.11577, -77.1894815))
}
