% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_maps_geocoding.R
\name{google_maps_geocode}
\alias{google_maps_geocode}
\title{Geocode an address.}
\usage{
google_maps_geocode(address_str = "900 N Glebe Rd, Arlington, VA 22203",
  api_key = "")
}
\arguments{
\item{address_str}{Street address string.}

\item{api_key}{Optional. Google geocoding API key. Needed if getting more than 2500 results in 1 day.}
}
\description{
Returns street_number, route, neighborhood, locality, administrative_area_level_2,
administrative_area_level_1, country, postal_code, lat, lng
}
\examples{
google_geocode("900 N Glebe Rd, Arlington, VA 22203", "AIzaSyC9FKW-kjQlEXjfM3OgMZBJ7xE6zCN1JQI")
}
