% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_maps_places.R
\name{google_maps_place_search}
\alias{google_maps_place_search}
\title{Search for places on Google Maps}
\usage{
google_maps_place_search(query, type = "", location = "", radius = 50000,
  api_key, nptk = "")
}
\arguments{
\item{query}{What you are searching for (e.g. "hotels near Arlington, VA").}

\item{type}{Optional. Search by Google Maps place type (e.g. "schools").}

\item{location}{Optional. The latitude/longitude around which to retrieve place
information. This must be specified as latitude,longitude.}

\item{radius}{Optional. Radius in meteres around location to search. Default is
50000 and the max.}

\item{api_key}{Google places API key.}
}
\description{
Returns up to 60 places and includes street_number, route, neighborhood, locality,
administrative_area_level_2, administrative_area_level_1, country, postal_code, lat, lng
}
\examples{
google_maps_place_search(query = "fast food in Fairfaz City VA", api_key = "AIzaSyC9FKW-kjQlEXjfM3OgMZBJ7xE6zCN1JQI")
}
